Add support for serialize => false and save_default => false
Unserialized values are saved on the root, meaning if you have a field named "foo" in a group "bar",
it gets saved under the meta value "foo" instead of "bar["foo"]".

If save_default is false any time the value == the default value, that meta row does not get saved (or gets
deleted if it already exists).